Tested and fixed lib.h

+++ Maths/LCM/LCM Pair/LCMPair.c
@@ -1,7 +1,6 @@
 #include <stdio.h>
-#include "../../GCD/GCD Pair/GCDPair.c"
 
 //Calculates the LCM of a and b.
-int lcm(int a,int b){
+long long int lcm(int a,int b){
 	return (a*b)/gcd(a,b);
 }
